pyproject.toml
[project]
name = "ai-agent-code-su"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
